backend/requirements.txt
flask
Flask-Cors
sqlalchemy
torch
pandas
scikit-learn
IPython
overrides
typeguard
transformers
seaborn
google-auth